refactor(bus): relocate Message/MessageKind into bus-core message module (Phase 3)
Move the canonical transport vocabulary out of microsvc::service into a
dependency-free transport::message module, so the bus stops returning a
microsvc error type from payload decoding:

- New transport/message.rs: Message, MessageKind, and a bus-core
  PayloadDecodeError. payload_json/payload_bitcode now return PayloadDecodeError
  instead of microsvc::HandlerError.
- microsvc::error gains From<PayloadDecodeError> for HandlerError; Context::input
  maps back, so handler signatures are unchanged.
- Re-export Message/MessageKind/PayloadDecodeError from both transport and
  microsvc, so every `crate::microsvc::Message` consumer is unaffected.

No behavior change. 229 lib + full default integration suite + 12 conformance
green; cargo check --all-features clean.

Implements [[tasks/bus-decomposition-phase1]] / [[specs/bus-module-decomposition]]
