+++ src/microsvc/service.rs
@@ -30,6 +30,7 @@ use super::context::Context;
 use super::dependencies::{HasReadModelStore, HasRepo, RepoReadModelDependencies};
 use super::error::HandlerError;
 use super::session::Session;
+use super::transport::{Message, MessageKind};
 
 type GuardFn<D> = dyn Fn(&Context<D>) -> bool + Send + Sync;
 type HandlerFuture<'a> = Pin<Box<dyn Future<Output = Result<Value, HandlerError>> + Send + 'a>>;
@@ -62,15 +63,6 @@ where
     F: for<'a> AsyncHandler<'a, D> + 'static,
 {
     Arc::new(move |ctx| Box::pin(handler.call(ctx)) as HandlerFuture<'_>)
-}
-
-/// The kind of message a handler consumes.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, serde::Deserialize, serde::Serialize)]
-pub enum MessageKind {
-    /// A command addressed to one handler.
-    Command,
-    /// A published event that may be consumed by many handlers.
-    Event,
 }
 
 /// How a handler expects the transport to deliver matching messages.
@@ -149,96 +141,6 @@ pub struct SubscriptionPlan {
     pub commands: Vec<String>,
     /// Event names consumed by fan-out event transports.
     pub events: Vec<String>,
-}
-
-/// Serializable transport message used by handlers.
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
-pub struct Message {
-    pub id: Option<String>,
-    pub name: String,
-    pub kind: MessageKind,
-    pub payload: Vec<u8>,
-    pub content_type: String,
-    pub metadata: Vec<(String, String)>,
-}
-
-impl Message {
-    /// Create a transport message.
-    pub fn new(name: impl Into<String>, kind: MessageKind, payload: Vec<u8>) -> Self {
-        Self {
-            id: None,
-            name: name.into(),
-            kind,
-            payload,
-            content_type: "application/json".to_string(),
-            metadata: Vec::new(),
-        }
-    }
-
-    /// Add a durable message id.
-    pub fn with_id(mut self, id: impl Into<String>) -> Self {
-        self.id = Some(id.into());
-        self
-    }
-
-    /// Add metadata.
-    pub fn with_metadata(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
-        self.metadata.push((key.into(), value.into()));
-        self
-    }
-
-    /// Get the durable message id, if this message has one.
-    pub fn id(&self) -> Option<&str> {
-        self.id.as_deref()
-    }
-
-    /// Get the message name.
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    /// Get the raw payload bytes.
-    pub fn payload(&self) -> &[u8] {
-        &self.payload
-    }
-
-    /// Get a metadata value by key.
-    pub fn metadata(&self, key: &str) -> Option<&str> {
-        self.metadata
-            .iter()
-            .find(|(existing, _)| existing.eq_ignore_ascii_case(key))
-            .map(|(_, value)| value.as_str())
-    }
-
-    /// Get the correlation id, if present.
-    pub fn correlation_id(&self) -> Option<&str> {
-        self.metadata("correlation_id")
-    }
-
-    /// Get the causation id, if present.
-    pub fn causation_id(&self) -> Option<&str> {
-        self.metadata("causation_id")
-    }
-
-    /// Decode the raw payload as JSON.
-    pub fn payload_json<T: serde::de::DeserializeOwned>(&self) -> Result<T, HandlerError> {
-        serde_json::from_slice(&self.payload).map_err(|e| {
-            HandlerError::DecodeFailed(format!(
-                "invalid JSON payload for message '{}': {}",
-                self.name, e
-            ))
-        })
-    }
-
-    /// Decode the raw payload as bitcode.
-    pub fn payload_bitcode<T: serde::de::DeserializeOwned>(&self) -> Result<T, HandlerError> {
-        bitcode::deserialize(&self.payload).map_err(|e| {
-            HandlerError::DecodeFailed(format!(
-                "invalid bitcode payload for message '{}': {}",
-                self.name, e
-            ))
-        })
-    }
 }
 
 /// A registered handler with optional guard.
