+++ src/microsvc/transport/mod.rs
@@ -79,14 +79,13 @@
 //! deduplication, when needed, is the optional consumer inbox, not an outbox or
 //! publish guarantee.
 
-use crate::microsvc::Message;
-
 mod bus;
 mod capabilities;
 mod error;
 mod failure_policy;
 mod handlers;
 mod in_memory_bus;
+mod message;
 #[cfg(feature = "kafka")]
 mod kafka;
 #[cfg(feature = "kafka")]
@@ -138,6 +137,7 @@ pub use error::{TransportError, TransportErrorKind};
 pub use failure_policy::{FailureAction, FailurePolicy};
 pub use handlers::{AsyncMessageHandler, Handlers};
 pub use in_memory_bus::{InMemoryBus, InMemoryReceived};
+pub use message::{Message, MessageKind, PayloadDecodeError};
 pub use outbox_dispatch::{OutboxDispatchOutcome, OutboxDispatcher, SOURCED_METADATA_PREFIX};
 pub use outbox_source::{
     OutboxSource, ReceivedOutboxMessage, DEFAULT_OUTBOX_SOURCE_BATCH, DEFAULT_OUTBOX_SOURCE_LEASE,
